tests/wpt/mozilla/tests/bluetooth/getDescriptors/get-same-descriptors.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}],
        optionalServices: [generic_access.name]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(generic_access.name))
    .then(service => service.getCharacteristic(device_name.name))
    .then(characteristic => Promise.all([
        characteristic.getDescriptors(number_of_digitals.name),
        characteristic.getDescriptors(number_of_digitals.name)]))
    .then(descriptors => assert_array_equals(descriptors[0], descriptors[1]));
}, 'Calls to get the same descriptor should return the same object.');
</script>
